LayoutTests/storage/indexeddb/intversion-close-in-upgradeneeded-expected.txt

Test that when db.close is called in upgradeneeded, the db is cleaned up on refresh.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "intversion-close-in-upgradeneeded.html"
indexedDB.deleteDatabase(dbname)
request = indexedDB.open(dbname, 7)


upgradeNeeded():
db = event.target.result
PASS event.newVersion is 7
transaction = event.target.transaction
db.createObjectStore('os')
db.close()

transaction.oncomplete:
sawTransactionComplete = true

openSuccess():
PASS sawTransactionComplete is true
db = event.target.result
PASS db.version is 7
transaction = db.transaction('os')
FAIL transaction = db.transaction('os') threw exception Error: InvalidStateError: DOM IDBDatabase Exception 11
PASS successfullyParsed is true

TEST COMPLETE
